backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.4.1
django-anymail==12.0
django-oauth-toolkit==3.0.1
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
